Clean-state pass: remove redundant export Python, dead code, rename onnx/ -> separation/
Architecture audit (3 agents) + cleanup:
- Remove tools/onnx_export/ entirely — the canonical export now lives in
  github.com/demixr/demucs-executorch (no longer duplicated here). Drop the
  dead .gitignore block for it.
- Rename lib/helpers/onnx/ -> lib/helpers/separation/: it hosts the ExecuTorch
  engine + engine-agnostic DSP (audio_io, demucs_config, istft), none ONNX-
  specific. Update all importers.
- Delete dead code: lib/components/animated_route.dart (orphan), DemucsConfig
  .sources6 + sourcesForCount (dead after 6-stem removal; use sources4).
- Remove the stale 6-stem integration test; fix device_check + onnx test for the
  new Model API (downloadUrl/onnxUrl).
- Rename the open_unmix.png asset -> demucs.png (+ call sites); retire the stale
  docs/ml-engine-migration.md (superseded by README); drop orphan
  ios/Runner/SceneDelegate.swift + stray logs.
- Format: dart format, swiftformat (ios/macos), google-java-format (MainActivity).
  Lint clean: flutter analyze, swiftlint. (Kept third_party/flutter_onnxruntime
  — still needed for the ORT graph-optimization OOM fix.)
