+++ android/app/src/main/java/com/demixr/demixr_app/MainActivity.java
@@ -2,5 +2,4 @@ package com.demixr.demixr_app;
 
 import io.flutter.embedding.android.FlutterActivity;
 
-public class MainActivity extends FlutterActivity {
-}
+public class MainActivity extends FlutterActivity {}
