+++ integration_test/gpu_vs_cpu_benchmark_test.dart
@@ -5,9 +5,9 @@ import 'package:integration_test/integration_test.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
 
-import 'package:demixr_app/helpers/onnx/demucs_config.dart';
-import 'package:demixr_app/helpers/onnx/executorch_demixing_engine.dart';
-import 'package:demixr_app/helpers/onnx/onnx_demixing_engine.dart';
+import 'package:demixr_app/helpers/separation/demucs_config.dart';
+import 'package:demixr_app/helpers/separation/executorch_demixing_engine.dart';
+import 'package:demixr_app/helpers/separation/onnx_demixing_engine.dart';
 
 /// GPU (ExecuTorch/CoreML) vs CPU (ONNX) benchmark on the same ~4-min input.
 ///
